Split Support.FakeMailer into BoomFakeMailer and SwooshFakeMailer
