+++ test/error_info_test.exs
@@ -36,7 +36,7 @@ defmodule ErrorInfoTest do
     use BoomNotifier,
       notifier: BoomNotifier.MailNotifier.Bamboo,
       options: [
-        mailer: Support.FakeMailer,
+        mailer: Support.BambooFakeMailer,
         from: "me@example.com",
         to: self(),
         subject: "BOOM error caught"
